fix(build): TIOCSCTTY ioctl cast wrong on aarch64
On aarch64, libc::ioctl's request argument is c_int, not c_ulong.
Removing the explicit cast and letting libc define the correct type
for each arch fixes the aarch64 musl release build.

+++ src/container.rs
@@ -3638,7 +3638,7 @@ impl Command {
                     libc::dup2(slave_fd, 0);
                     libc::dup2(slave_fd, 1);
                     libc::dup2(slave_fd, 2);
-                    libc::ioctl(slave_fd, libc::TIOCSCTTY as libc::c_ulong, 0i32);
+                    libc::ioctl(slave_fd, libc::TIOCSCTTY, 0);
                     if slave_fd > 2 {
                         libc::close(slave_fd);
                     }
@@ -5376,7 +5376,7 @@ impl Command {
                     libc::dup2(slave_fd, 0);
                     libc::dup2(slave_fd, 1);
                     libc::dup2(slave_fd, 2);
-                    libc::ioctl(slave_fd, libc::TIOCSCTTY as libc::c_ulong, 0i32);
+                    libc::ioctl(slave_fd, libc::TIOCSCTTY, 0);
                     if slave_fd > 2 {
                         libc::close(slave_fd);
                     }
